server/industry-classifier-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
transformers>=4.0.0
torch>=1.8.0 # Or tensorflow>=2.0.0 if preferred
# Add other specific libraries later if needed
